analysis/requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter environment
jupyter>=1.0.0
ipykernel>=6.22.0

# Machine learning
scikit-learn>=1.2.0

# Audio processing
librosa>=0.10.0

# Environment variables
python-dotenv>=1.0.0

# Development tools
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
ruff>=0.0.262

# Type stubs
pandas-stubs>=2.0.0.0
types-requests>=2.28.11.0

# Optional: Natural language processing (uncomment if needed)
# nltk>=3.8.1
# spacy>=3.5.2
# transformers>=4.28.1

# Optional: Deep learning (uncomment if needed)
# torch>=2.0.0
# tensorflow>=2.12.0 